+++ test.js
@@ -3,14 +3,15 @@ const erre = require('./erre')
 
 const add = (a, b) => a + b
 const multiply = (a, b) => a * b
+const delay = (timeout) => new Promise(resolve => setTimeout(resolve, timeout))
 
 describe('erre', () => {
   it('can listen and dispatch simple events', (done) => {
     const stream = erre()
     const startValue = 1
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         assert.equal(startValue, value)
         done()
       })
@@ -23,15 +24,15 @@ describe('erre', () => {
     const eventsAmount = 4
     let currentEventId = 0
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         currentEventId ++
         assert.equal(startValue, value)
         if (currentEventId === eventsAmount) done()
       })
       .push(startValue)
-      .push(startValue)
-      .push(startValue)
+      .push(delay(100).then(() => startValue))
+      .push(delay(40).then(() => startValue))
       .push(startValue)
   })
 
@@ -43,8 +44,8 @@ describe('erre', () => {
 
     const expectedValue = 3
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         assert.equal(expectedValue, value)
         done()
       })
@@ -59,8 +60,8 @@ describe('erre', () => {
 
     const startValue = 1
 
-    stream
-      .onError((value) => {
+    stream.on
+      .error((value) => {
         done()
       })
       .push(startValue)
@@ -77,8 +78,8 @@ describe('erre', () => {
 
     const startValue = 1
 
-    stream
-      .onError((value) => {
+    stream.on
+      .error((value) => {
         done()
       })
       .push(startValue)
@@ -89,8 +90,8 @@ describe('erre', () => {
     const startValue = 1
     const expectedValue = 2
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         assert.equal(value, expectedValue)
         done()
       }).push(startValue)
@@ -108,8 +109,8 @@ describe('erre', () => {
     const startValue = 1
     const expectedValue = 2
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         assert.equal(value, expectedValue)
         done()
       })
@@ -120,13 +121,26 @@ describe('erre', () => {
     const stream = erre()
     const startValue = 1
 
-    stream.onEnd(() => done())
+    stream.on.end(() => done())
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         throw 'it should never be called if the stream was ended'
       })
       .end()
+      .push(startValue)
+  })
+
+  it('can be ended avoiding race condition issues', (done) => {
+    const stream = erre()
+    const startValue = 1
+
+    stream.on.end(() => done())
+
+    stream.on
+      .value((value) => {
+        stream.end()
+      })
       .push(startValue)
   })
 
@@ -137,8 +151,8 @@ describe('erre', () => {
 
     stream.connect(val => multiply(val, 2))
 
-    stream
-      .onValue((value) => {
+    stream.on
+      .value((value) => {
         assert.equal(value, expectedValue)
         done()
       })
@@ -153,11 +167,42 @@ describe('erre', () => {
 
     stream.connect(val => multiply(val, 2))
 
-    fork
-      .onValue((value) => {
+    fork.on
+      .value((value) => {
         assert.equal(value, expectedValue)
         done()
       })
       .push(startValue)
+  })
+
+  it('stream.next will not dispatch events', () => {
+    const stream = erre((val) => add(val, 1))
+    const startValue = 1
+    const expectedValue = 2
+
+    stream.on.value(() => {
+      throw 'no events should be dispatched'
+    })
+
+    const { done, value } = stream.next(startValue)
+
+    return value.then((val) => {
+      assert.equal(val, expectedValue)
+      assert.equal(done, false)
+    })
+  })
+
+  it('stream.next will have undefined values if ended', () => {
+    const stream = erre()
+    const expectedValue = 2
+
+    stream.end()
+
+    const { done, value } = stream.next(() => {
+      return expectedValue
+    })
+
+    assert.equal(done, true)
+    assert.equal(value, undefined)
   })
 })