+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # erre
 
+## 2.0.0
+
+- changed: all the `on*` API methods were linked to the `on` object for example: `onValue` => `on.value`
+- added: the `next` method to run raw stream events without dispatching them to the listeners
+
 ## 1.2.0
 
 - updated: return the stream generator extended with erre API methods